+++ lib/money/exchange_rates/open_exchange_rates.ex
@@ -1,10 +1,11 @@
 defmodule Money.ExchangeRates.OpenExchangeRates do
   @behaviour Money.ExchangeRates
 
-  @app_id Money.get_env(:open_exchange_rates_app_id)
+  @app_id Money.get_env(:open_exchange_rates_app_id, "")
+  @api_module Money.get_env(:api_module, __MODULE__)
   @exr_url "https://openexchangerates.org/api"
 
-  if is_nil(@app_id) do
+  if @app_id == "" and @api_module == __MODULE__ do
     raise ArgumentError, message: "An Open Exchange Rates app_id must be configured in config.exs"
   end